Refactor tfno.py -> fno.py
