+++ neuralop/models/__init__.py
@@ -1,5 +1,5 @@
-from .tfno import TFNO, TFNO1d, TFNO2d, TFNO3d
-from .tfno import FNO, FNO1d, FNO2d, FNO3d
-from .tfno import SFNO
+from .fno import TFNO, TFNO1d, TFNO2d, TFNO3d
+from .fno import FNO, FNO1d, FNO2d, FNO3d
+from .fno import SFNO
 from .uno import UNO
 from .model_dispatcher import get_model
